Migrate core client to MainApp
